+++ Lab Assignment-1/main_omp.c
@@ -31,7 +31,7 @@ int main(int argc, char const *argv[])
 	int num_threads;		//no. of threads to be used (input)
 	int* data_points;		//data points (input)
 	int* cluster_points;	//clustered data points (to be computed)
-	int* centroids;			//centroids of each iteration (to be computed)
+	float* centroids;			//centroids of each iteration (to be computed)
 	int num_iterations;    //no of iterations performed by algo (to be computed)
 	//---------------------------------------------------------------------
